Refactor coalescing logic out of the file context.
